intelli-credit/ai-services/cam-generator/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
python-docx==1.1.0
python-dateutil==2.8.2
